Refactor TechnologiesInfo
